+++ TaxiDetails/TaxiDetails.Domain.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34728.123
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaxiDetails", "TaxiDetails\TaxiDetails.csproj", "{255199C4-E223-4360-8522-C6A7A435E4F4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{255199C4-E223-4360-8522-C6A7A435E4F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{255199C4-E223-4360-8522-C6A7A435E4F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{255199C4-E223-4360-8522-C6A7A435E4F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{255199C4-E223-4360-8522-C6A7A435E4F4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E33F2963-C73E-41B0-8078-C05B66A2AF36}
+	EndGlobalSection
+EndGlobal
