--- TaxiDetails/TaxiDetails.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaxiDetails", "TaxiDetails\TaxiDetails.csproj", "{94FC8B83-0376-4ABE-AA47-052917FF57C8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaxiDetailsClass", "TaxiDetailsClass\TaxiDetailsClass.csproj", "{09C41C82-5D29-4621-A115-F61A9425B605}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaxiDetails.Tests", "TaxiDetails.Tests\TaxiDetails.Tests.csproj", "{06D3B5BD-DE4D-4773-887F-528C708A28B8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{94FC8B83-0376-4ABE-AA47-052917FF57C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94FC8B83-0376-4ABE-AA47-052917FF57C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94FC8B83-0376-4ABE-AA47-052917FF57C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94FC8B83-0376-4ABE-AA47-052917FF57C8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{09C41C82-5D29-4621-A115-F61A9425B605}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09C41C82-5D29-4621-A115-F61A9425B605}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{09C41C82-5D29-4621-A115-F61A9425B605}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{09C41C82-5D29-4621-A115-F61A9425B605}.Release|Any CPU.Build.0 = Release|Any CPU
-		{06D3B5BD-DE4D-4773-887F-528C708A28B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{06D3B5BD-DE4D-4773-887F-528C708A28B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{06D3B5BD-DE4D-4773-887F-528C708A28B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{06D3B5BD-DE4D-4773-887F-528C708A28B8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2B71969F-12D2-466D-A438-D813A09733D2}
-	EndGlobalSection
-EndGlobal
